src/main/java/com/hostfully/web/rest/BookingController.java
package com.hostfully.web.rest;

import com.hostfully.entity.Booking;
import com.hostfully.service.BookingService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/v1/bookings")
public class BookingController {

    @Autowired
    private BookingService bookingService;

    @PostMapping
    public ResponseEntity<Booking> createBooking(@RequestBody Booking booking) {
        Booking createdBooking = bookingService.createBooking(booking);
        return ResponseEntity.ok(createdBooking);
    }

    @PutMapping("/{id}")
    public ResponseEntity<Booking> updateBooking(@PathVariable Long id, @RequestBody Booking booking) {
        Booking updatedBooking = bookingService.updateBooking(id, booking);
        return ResponseEntity.ok(updatedBooking);
    }

    @PutMapping("/{id}/cancel")
    public ResponseEntity<Void> cancelBooking(@PathVariable Long id) {
        bookingService.cancelBooking(id);
        return ResponseEntity.noContent().build();
    }

    @PutMapping("/{id}/rebook")
    public ResponseEntity<Booking> rebook(@PathVariable Long id) {
        Booking rebooked = bookingService.rebook(id);
        return ResponseEntity.ok(rebooked);
    }

    @GetMapping("/{id}")
    public ResponseEntity<Booking> getBooking(@PathVariable Long id) {
        Booking booking = bookingService.getBooking(id);
        return ResponseEntity.ok(booking);
    }

    @GetMapping
    public ResponseEntity<List<Booking>> getAllBooking() {
       List<Booking> booking = bookingService.getAllBookings();
        return ResponseEntity.ok(booking);
    }

    @GetMapping("/person/{personId}")
    public ResponseEntity<List<Booking>> getBookingsByPerson(@PathVariable Long personId) {
        List<Booking> bookings = bookingService.findBookingsByPersonId(personId);
        if (bookings.isEmpty()) {
            return ResponseEntity.noContent().build();
        }
        return ResponseEntity.ok(bookings);
    }
}
